Se elimina slider para ahorrar código y se unas mainCard como remplazo, se añade una variable que mete las mainCards a mainCard
